+++ support-modules/supporter-product-data-dynamo/src/main/scala/com/gu/supporterdata/model/FieldNames.scala
@@ -1,5 +1,4 @@
-package com.gu.model.dynamo
-
+package com.gu.supporterdata.model
 
 object FieldNames {
   val subscriptionName = "subscriptionName"
