+++ supporter-product-data/src/test/scala/com/gu/lambdas/UpdateDynamoSpec.scala
@@ -1,12 +1,12 @@
 package com.gu.lambdas
 
 import com.gu.Fixtures
-import com.gu.model.dynamo.SupporterRatePlanItem
 import kantan.csv.ops.toCsvInputOps
 import kantan.csv.rfc
 import org.scalatest.flatspec.AsyncFlatSpec
 import org.scalatest.matchers.should.Matchers
 import com.gu.model.dynamo.SupporterRatePlanItemCodecs._
+import com.gu.supporterdata.model.SupporterRatePlanItem
 
 class UpdateDynamoSpec extends AsyncFlatSpec with Matchers {
