+++ supporter-product-data/src/test/scala/com/gu/services/DynamoDBServiceSpec.scala
@@ -1,7 +1,9 @@
 package com.gu.services
 
+import com.gu.{model, supporterdata}
 import com.gu.model.Stage.DEV
-import com.gu.model.dynamo.SupporterRatePlanItem
+import com.gu.supporterdata.model.SupporterRatePlanItem
+import com.gu.supporterdata.services.SupporterDataDynamoService
 import com.gu.test.tags.annotations.IntegrationTest
 import org.scalatest.flatspec.AsyncFlatSpec
 import org.scalatest.matchers.should.Matchers
@@ -13,7 +15,7 @@ class DynamoDBServiceSpec  extends AsyncFlatSpec with Matchers {
 
   SupporterDataDynamoService.getClass.getSimpleName should "be able to insert an item" in {
     val service = SupporterDataDynamoService(DEV)
-    val item = SupporterRatePlanItem(
+    val item = supporterdata.model.SupporterRatePlanItem(
       identityId = "999999",
       gifteeIdentityId = None,
       subscriptionName = "test-sub-name",
